@@ -25,6 +25,7 @@
     "@types/lodash.throttle": "^4.1.3",
     "@types/markdown-it": "0.0.2",
     "@types/moment": "^2.13.0",
+    "@types/prop-types": "^15.5.2",
     "@types/react-portal": "^3.0.1",
     "@types/validator": "^6.2.0",
     "aphrodite": "^1.2.0",
@@ -37,18 +38,19 @@
     "markdown-it": "^8.3.1",
     "markdown-it-katex": "^2.0.3",
     "moment": "^2.18.1",
-    "next": "^2.0.1",
-    "react": "^15.4.2",
-    "react-dom": "^15.4.2",
+    "next": "^4.0",
+    "prop-types": "^15.6.0",
+    "react": "^16.2",
+    "react-dom": "^16.2",
     "react-measure": "^1.4.7",
     "react-numeric-input": "^2.0.7",
     "react-pointable": "^0.1.0",
     "react-portal": "^3.0.0",
-    "react-radio-group": "^3.0.1",
+    "react-radio-group": "marudor/react-radio-group",
     "react-visibility-sensor": "^3.10.1",
-    "slate": "^0.25",
-    "slate-plain-serializer": "^0.1.11",
-    "slate-react": "^0.2.1",
+    "slate": "^0.32",
+    "slate-plain-serializer": "^0.5",
+    "slate-react": "^0.11",
     "validator": "^7.2.0"
   },
   "devDependencies": {
@@ -65,7 +67,6 @@
     "lint-staged": "^3.4.0",
     "now": "^4.10.3",
     "prettier": "1.7.2",
-    "react-no-ssr": "^1.1.0",
     "rimraf": "^2.6.1"
   },
   "now": {
