@@ -11,14 +11,14 @@
 
 const React = require("react");
 const { StyleSheet, css } = require("aphrodite");
-const { PropTypes } = React;
+import PropTypes from "prop-types";
 
 const globalStyles = require("../styles.js");
 const { colors } = globalStyles;
 
-const childrenType = React.PropTypes.oneOfType([
-  React.PropTypes.arrayOf(React.PropTypes.node),
-  React.PropTypes.node,
+const childrenType = PropTypes.oneOfType([
+  PropTypes.arrayOf(PropTypes.node),
+  PropTypes.node,
 ]);
 
 class Button extends React.Component {
@@ -172,11 +172,7 @@ class Button extends React.Component {
       ...otherProps
     } = this.props;
 
-    const {
-      hideFocus,
-      pressed,
-      hover,
-    } = this.state;
+    const { hideFocus, pressed, hover } = this.state;
 
     const styleOverrides = pressed
       ? pressStyle || style
