@@ -4,7 +4,7 @@ import Link from "next/link";
 import React from "react";
 import { StyleSheet, css } from "aphrodite";
 import Plain from "slate-plain-serializer";
-import { State } from "slate";
+import { Value } from "slate";
 
 import Button from "../button";
 import PageButton from "../page-button";
@@ -17,7 +17,7 @@ const validateData = data => {
     return false;
   } else if (typeof data === "object" && data.kind === "rich-editor") {
     // TODO(andy): extract this to rich-editor or a helper module; duplicative.
-    return Plain.serialize(State.fromJSON(JSON.parse(data.rawData))).length > 0;
+    return Plain.serialize(Value.fromJSON(JSON.parse(data.rawData))).length > 0;
   } else if (typeof data === "object" && data.kind === "scratchpad") {
     // TODO(andy): extract this to scratchpad-canvas or a helper module; duplicative.
     return data.totalInkLength > 50;
