@@ -20,10 +20,11 @@ const styles = StyleSheet.create({
     height: "100%",
     minHeight: "100%",
     boxSizing: "border-box",
-    backgroundColor: sharedStyles.colors.gray90,
     [mediaQueries.lgOrLarger]: {
       justifyContent: "center",
     },
+    "-webkit-overflow-scrolling": "touch",
+    overflow: "auto",
   },
 
   innerContainer: {
