@@ -18,7 +18,14 @@ const ResponseCard = (props: Props) => {
         <div className={css(styles.avatar)} />
         <p className={css(styles.studentName)}>{props.studentName}</p>
       </div>
-      <RichEditor placeholder="Write your response here." editable />
+      <RichEditor
+        placeholder="Type your response here."
+        onChange={value => {
+          props.onChange(value);
+        }}
+        data={props.data}
+        editable
+      />
       <div className={css(styles.footer)}>
         <div className={css(styles.hairline)} />
         <Button style={styles.submitButton}>Submit Response</Button>
