@@ -36,7 +36,6 @@ export default class RichEditor extends React.Component {
 
   constructor(props: Props) {
     super(props);
-    Slate.resetKeyGenerator();
     this.state = {
       state: this.getEditorStateFromProps(props),
       menu: null,
@@ -55,60 +54,60 @@ export default class RichEditor extends React.Component {
           </span>
         ),
       },
-      rules: [
-        {
-          match: object => object.kind === "text",
-          decorate: (text, node) => {
-            if (!this.props.diffBase) {
-              return;
+    };
+
+    this.plugins = [
+      {
+        decorateNode: (text, node) => {
+          if (!this.props.diffBase) {
+            return;
+          }
+
+          const baseDocument = Slate.Value.fromJSON(
+            JSON.parse(this.props.diffBase.rawData),
+          ).document;
+          const currentDocument = this.state.state.document;
+          const diffs = diffWordsWithSpace(
+            baseDocument.text,
+            currentDocument.text,
+          );
+
+          let characters = text.characters.asMutable();
+
+          let index = 0;
+          const document = this.state.state.document;
+          for (let diff of diffs) {
+            if (diff.removed) {
+              continue;
             }
-
-            const baseDocument = Slate.State.fromJSON(
-              JSON.parse(this.props.diffBase.rawData),
-            ).document;
-            const currentDocument = this.state.state.document;
-            const diffs = diffWordsWithSpace(
-              baseDocument.text,
-              currentDocument.text,
-            );
-
-            let characters = text.characters.asMutable();
-
-            let index = 0;
-            const document = this.state.state.document;
-            for (let diff of diffs) {
-              if (diff.removed) {
-                continue;
-              }
-              if (diff.added) {
-                for (
-                  let characterIndex = index;
-                  characterIndex < index + diff.value.length;
-                  characterIndex++
-                ) {
-                  const characterText = currentDocument.getTextAtOffset(
-                    characterIndex,
-                  );
-                  if (characterText === text) {
-                    const textNodeOffset = document.getOffset(text.key);
-                    const textCharacterIndex = characterIndex - textNodeOffset;
-                    let character = characters.get(textCharacterIndex);
-                    let { marks } = character;
-                    let mark = Mark.create({ type: "changed" });
-                    marks = marks.add(mark);
-                    character = character.merge({ marks });
-                    characters.set(textCharacterIndex, character);
-                  }
+            if (diff.added) {
+              for (
+                let characterIndex = index;
+                characterIndex < index + diff.value.length;
+                characterIndex++
+              ) {
+                const characterText = currentDocument.getTextAtOffset(
+                  characterIndex,
+                );
+                if (characterText === text) {
+                  const textNodeOffset = document.getOffset(text.key);
+                  const textCharacterIndex = characterIndex - textNodeOffset;
+                  let character = characters.get(textCharacterIndex);
+                  let { marks } = character;
+                  let mark = Mark.create({ type: "changed" });
+                  marks = marks.add(mark);
+                  character = character.merge({ marks });
+                  characters.set(textCharacterIndex, character);
                 }
               }
-              index += diff.value.length;
             }
+            index += diff.value.length;
+          }
 
-            return characters.asImmutable();
-          },
+          return characters.asImmutable();
         },
-      ],
-    };
+      },
+    ];
   }
 
   componentWillReceiveProps = (nextProps: Props) => {
@@ -135,7 +134,7 @@ export default class RichEditor extends React.Component {
     if (kind !== dataKind) {
       console.error(`Unexpected data kind in ${effectiveProps.data}`);
     }
-    return Slate.State.fromJSON(JSON.parse(rawData));
+    return Slate.Value.fromJSON(JSON.parse(rawData));
   };
 
   componentDidUpdate = () => {
@@ -149,10 +148,10 @@ export default class RichEditor extends React.Component {
 
   // TODO(andy): Wrap Slate's type.
   onChange = (change: any) => {
-    const { state } = change;
+    const { value } = change;
 
-    this.setState({ state });
-    const serializedState = state.toJSON();
+    this.setState({ state: value });
+    const serializedState = value.toJSON();
     const stringifiedState = JSON.stringify(serializedState);
     const stringifiedCurrentState = JSON.stringify(this.state.state.toJSON());
     // State changes when e.g. selection changes, but we don't want to bother passing that along.
@@ -266,11 +265,12 @@ export default class RichEditor extends React.Component {
       >
         {this.renderMenu()}
         <Editor
+          onChange={this.onChange}
+          value={editorState}
           schema={this.schema}
-          state={editorState}
+          plugins={this.plugins}
           onFocus={this.onFocus}
           onBlur={this.onBlur}
-          onChange={this.onChange}
           onKeyDown={this.onKeyDown}
           readOnly={!this.props.editable || this.props.quotedWork}
           placeholder={
