@@ -2,7 +2,7 @@
 import { StyleSheet, css } from "aphrodite";
 import MarkdownIt from "markdown-it";
 import MarkdownItKaTeX from "markdown-it-katex";
-import { State } from "slate";
+import { Value } from "slate";
 import Plain from "slate-plain-serializer";
 
 import sharedStyles from "../styles.js";
@@ -21,7 +21,7 @@ export default (props: {
     return null;
   } else if (typeof contents === "object" && contents.kind === "rich-editor") {
     // TODO(andy): extract this to rich-editor or a helper module; duplicative.
-    const slateData = State.fromJSON(JSON.parse(contents.rawData), {
+    const slateData = Value.fromJSON(JSON.parse(contents.rawData), {
       terse: true,
     });
     contents = Plain.serialize(slateData);
